make ssr package work for Remix and Next.js
