+++ packages/ssr/README.md
@@ -2,7 +2,7 @@
 
 > A package for server-side rendering Stencil components.
 
-This package provides utilities for server-side rendering Stencil components in React, Vue or Angular applications bundled through a Rollup based compiler, e.g. Vite or Remix.
+This package provides utilities for server-side rendering Stencil components in React. You can use it in Next.js or Vite based applications, e.g. Remix.
 
 ## Install
 
@@ -18,7 +18,7 @@ In your e.g. Vite configuration, add the `stencilSSR` plugin as following:
 
 ```ts
 import { defineConfig } from "vite";
-import { stencilSSR } from "@stencil/react-output-target/vite";
+import { stencilSSR } from "@stencil/ssr";
 
 export default defineConfig({
   // ...
@@ -27,21 +27,46 @@ export default defineConfig({
       module: import('component-library-react'),
       from: 'component-library-react',
       hydrateModule: import('component-library/hydrate'),
+      serializeShadowRoot: 'declarative-shadow-dom'
     }),
   ],
 });
 ```
 
+### Next.js
+
+If you are using Next.js, you can wrap your `NextConfig` as following:
+
+```ts
+// next.conf.mjs
+import stencilSSR from '@stencil/ssr/next';
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {};
+
+export default stencilSSR({
+  module: import('component-library-react'),
+  from: 'component-library-react',
+  hydrateModule: import('component-library/hydrate'),
+  serializeShadowRoot: {
+    'scoped': ['my-counter'],
+    default: 'declarative-shadow-dom',
+  },
+})(nextConfig);
+```
+
+## Options
+
 The plugin requires the following options:
 
 #### `module`
 
-The import of the library that exposes your Stencil components wrapped around the output target.
+The package or module import that exposes your wrapped Stencil components using the [React Output Target](https://www.npmjs.com/package/@stencil/react-output-target).
 
 ### `from`
 
-The name of the library that exposes your Stencil components wrapped around the output target.
+The name of the library that exposes your Stencil components.
 
 ### `hydrateModule`
 
-tbd.
+The import of the hydrate module generated by the [Stencil hydrate output target](https://stenciljs.com/docs/hydrate-app#how-to-use-the-hydrate-app).
