+++ packages/ssr/package.json
@@ -19,8 +19,11 @@
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
+      "import": "./dist/index.js"
+    },
+    "./next": {
+      "types": "./dist/next.d.ts",
+      "import": "./dist/next.js"
     }
   },
   "files": [
@@ -42,17 +45,19 @@
   },
   "devDependencies": {
     "@types/node": "^20.14.12",
-    "typescript": "^5.4.4"
+    "typescript": "^5.4.4",
+    "webpack": "^5.98.0"
   },
   "dependencies": {
     "ast-types": "^0.14.2",
     "decamelize": "^6.0.0",
     "esbuild": "^0.25.0",
+    "imports-loader": "^5.0.0",
     "mlly": "^1.7.4",
-    "recast": "^0.23.9",
-    "ts-morph": "^22.0.0"
+    "recast": "^0.23.9"
   },
   "peerDependencies": {
-    "vite": "^6.x"
+    "vite": "^6.x",
+    "next": "13 || 14 || 15"
   }
 }
