+++ packages/ssr/vite.config.ts
@@ -5,10 +5,8 @@ export default defineConfig({
   build: {
     lib: {
       entry: {
-        'react-output-target': 'src/react-output-target/index.ts',
-        react: 'src/react/index.ts',
-        ssr: 'src/react/ssr.tsx',
-        vite: 'src/react/vite.ts',
+        index: 'src/index.ts',
+        next: 'src/next.ts'
       },
       formats: ['es', 'cjs'],
     },
@@ -19,14 +17,7 @@ export default defineConfig({
         'node:path', 'node:url', 'node:fs',
         '@stencil/core', '@lit/react', 'typescript', 'react', 'react-dom/server', 'ts-morph',
         'html-react-parser', 'mlly', 'esbuild', 'recast', 'ast-types',
-      ],
-      output: {
-        // Provide global variables to use in the UMD build
-        // for externalized deps
-        globals: {
-          'ts-morph': 'tsMorph',
-        },
-      },
+      ]
     },
   },
 });
